backend/requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
cffi==1.17.1
cryptography==44.0.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnspython==2.7.0
paramiko==3.5.1
psycopg2==2.9.10
pycparser==2.22
pymongo==4.11.1
PyNaCl==1.5.0
python-dotenv==1.0.1
sqlparse==0.5.3
sshtunnel==0.4.0
